split-service/requirements.txt
flask
opencv-python
requests
redis
google-cloud-storage
pika
